Extract logic for generating Refresh and Access Tokens and setting them on the HTTP response to separate services
