+++ application/account-management/Application/ApplicationConfiguration.cs
@@ -1,6 +1,8 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using PlatformPlatform.AccountManagement.Application.Authentication;
 using PlatformPlatform.SharedKernel.ApplicationCore;
+using PlatformPlatform.SharedKernel.ApplicationCore.Authentication;
 
 namespace PlatformPlatform.AccountManagement.Application;
 
@@ -17,6 +19,10 @@ public static class ApplicationConfiguration
         var securityTokenSettings = configuration.GetSection("SecurityTokenSettings").Get<SecurityTokenSettings>()
                                     ?? throw new InvalidOperationException("No SecurityTokenSettings configuration found.");
         services.AddSingleton(securityTokenSettings);
+
+        services.AddSingleton<SecurityTokenGenerator>();
+        services.AddTransient<SecurityTokenService>();
+
         return services;
     }
 }
