+++ application/account-management/Application/Authentication/CompleteLogin.cs
@@ -1,8 +1,3 @@
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text.Json;
-using Microsoft.AspNetCore.Http;
-using Microsoft.IdentityModel.Tokens;
 using PlatformPlatform.AccountManagement.Application.TelemetryEvents;
 using PlatformPlatform.AccountManagement.Domain.Authentication;
 using PlatformPlatform.SharedKernel.ApplicationCore.Authentication;
@@ -22,14 +17,11 @@ public sealed class CompleteLoginHandler(
     IUserRepository userRepository,
     ILoginRepository loginProcessRepository,
     OneTimePasswordValidator oneTimePasswordValidator,
+    SecurityTokenService securityTokenService,
     ITelemetryEventsCollector events,
-    IHttpContextAccessor httpContextAccessor,
     ILogger<CompleteLoginHandler> logger
 ) : IRequestHandler<CompleteLoginCommand, Result>
 {
-    // TODO: Change this to use ASP.NET data protection API and revert this commit from history before merging
-    private static readonly byte[] Key = "put-64-bytes-key-here"u8.ToArray();
-
     public async Task<Result> Handle(CompleteLoginCommand command, CancellationToken cancellationToken)
     {
         var loginProcess = await loginProcessRepository.GetByIdAsync(command.Id, cancellationToken);
@@ -75,52 +67,10 @@ public sealed class CompleteLoginHandler(
         loginProcess.MarkAsCompleted();
         loginProcessRepository.Update(loginProcess);
 
-        CreateAuthenticationTokens(user);
+        securityTokenService.CreateAndSetSecurityTokens(user);
 
         events.CollectEvent(new LoginCompleted(user.Id, (int)registrationTimeInSeconds));
 
         return Result.Success();
-    }
-
-    private void CreateAuthenticationTokens(User user)
-    {
-        var httpContext = httpContextAccessor.HttpContext ?? throw new InvalidOperationException("HttpContext is null.");
-
-        const string issuer = "https://localhost:9000";
-        const string audience = "https://localhost:9000";
-
-        var tokenDescriptor = new SecurityTokenDescriptor
-        {
-            Subject = new ClaimsIdentity(new[]
-                {
-                    new Claim("Id", Guid.NewGuid().ToString()),
-                    new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
-                    new Claim(JwtRegisteredClaimNames.Sub, user.Id),
-                    new Claim(JwtRegisteredClaimNames.Email, user.Email),
-                    new Claim(JwtRegisteredClaimNames.GivenName, user.FirstName ?? string.Empty),
-                    new Claim(JwtRegisteredClaimNames.FamilyName, user.LastName ?? string.Empty),
-                    new Claim("tenantId", user.TenantId),
-                    new Claim("role", user.Role.ToString()),
-                    new Claim("locale", "en"),
-                    new Claim("title", user.Title ?? string.Empty),
-                    new Claim("avatarUrl", user.Avatar.Url ?? string.Empty)
-                }
-            ),
-            Expires = DateTime.UtcNow.AddMinutes(5),
-            Issuer = issuer,
-            Audience = audience,
-            SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(Key), SecurityAlgorithms.HmacSha512Signature)
-        };
-
-        var refreshToken = new RefreshToken { UserId = user.Id };
-        httpContext.Response.Headers.Remove(RefreshToken.XRefreshTokenKey);
-        httpContext.Response.Headers.Append(RefreshToken.XRefreshTokenKey, JsonSerializer.Serialize(refreshToken));
-
-        var tokenHandler = new JwtSecurityTokenHandler();
-        var securityToken = tokenHandler.CreateToken(tokenDescriptor);
-        var accessToken = tokenHandler.WriteToken(securityToken);
-
-        httpContext.Response.Headers.Remove(RefreshToken.XAccessTokenKey);
-        httpContext.Response.Headers.Append(RefreshToken.XAccessTokenKey, accessToken);
     }
 }
